dailyemail/pyproject.toml
[project]
name = "dailyemail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "dotenv>=0.9.9",
    "schedule>=1.2.2",
]
